requirement.txt
kafka-python 
matplotlib
re
fast-api
confluent-kafka